requirements.txt
Django==1.11
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-bulk==0.2.1
Pygments==2.2.0
pytz==2018.3
suds-jurko==0.6
